requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
celery[redis]==5.3.6
celery-redbeat==2.2.0
qdrant-client==1.7.0
sentence-transformers==2.3.1
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.10.0
pydantic==2.5.3
PyYAML==6.0.1
httpx==0.26.0
faster-whisper==0.10.0
pyttsx3==2.90
python-multipart==0.0.6
watchdog==3.0.0
pypdf==4.0.0
python-docx==1.1.0
sse-starlette==2.0.0
aiofiles==23.2.1
